Move more fragments into their own package
